Conflicts:
	C/c_interface.c
	C/exec.c

Former-commit-id: 1e379e7635a5f4c9a223c17b1c4da3bc4ef63498

+++ OPTYap/opt.preds.c
@@ -120,7 +120,7 @@ struct page_statistics {
 #ifdef USE_PAGES_MALLOC
   long pages_in_use;   /* same as struct pages (opt.structs.h) */
 #endif                 /* USE_PAGES_MALLOC */
-  long structs_in_use; /* same as struct pages (opt.structs.h) */
+  size_t structs_in_use; /* same as struct pages (opt.structs.h) */
   long bytes_in_use;
 };
